third_party/WebKit/LayoutTests/virtual/serviceworker/http/tests/serviceworker/current-on-load-expected.txt



PASS current is set for a controlled document 
